refactor: Extract queue processing logic to separate module
Markdown-generator main.py was 512 lines - too large and complex.

Changes:
- Created queue_processor.py with queue processing logic
- Extracted create_message_handler() function
- Extracted startup_queue_processor() function
- Extracted signal_site_publisher() helper function
- main.py now focused on FastAPI routes and app lifecycle

Results:
- main.py: 512 to 391 lines (121 lines removed)
- queue_processor.py: 186 lines (new module)
- Better separation of concerns
- Easier to test and maintain
- All 25 tests still passing
